Split diffsol initial-value projection into its own module
The solver-selection commit pushed rumoca-solver-diffsol/src/lib.rs to
2028 lines, tripping the SPEC_0021 2000-line hard limit — the root
cause of every CI failure on #191 (review-scan forbidden finding in the
lint job and test_rs_files_stay_under_spec_0021_hard_limit on all three
OSes).

Move the initial-value projection block (initial event/pre settling,
algebraic projection, initialization-plan unknowns, plus the
EventObservation recorder and the initial-event flag helper it owns)
to init_projection.rs: lib.rs 2028 -> 1697 lines with the cohesive
block intact rather than an arbitrary cut.
